feat: add remaining dtos

+++ src/main/java/br/com/fiap/chameleonfutureacademy/service/User/UserServiceImpl.java
@@ -2,8 +2,10 @@ package br.com.fiap.chameleonfutureacademy.service.User;
 
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.com.fiap.chameleonfutureacademy.domainmodel.User;
 import br.com.fiap.chameleonfutureacademy.domainmodel.exceptions.FieldValidationException;
@@ -35,6 +37,9 @@ public class UserServiceImpl implements UserService<User, Long> {
 
     @Override
     public User update(User user) {
+        if (existsById(user.getUserId()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário não encontrado");
+
         userRepository.findByEmail(user.getEmail()).ifPresent(other -> {
             if (!other.getUserId().equals(user.getUserId())) {
                 throw new FieldValidationException("email", "Este e-mail já está em uso.");
@@ -46,6 +51,9 @@ public class UserServiceImpl implements UserService<User, Long> {
 
     @Override
     public void removeById(Long id) {
+        if (existsById(id))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário não encontrado");
+
         userRepository.deleteById(id);
     }
